Lift state from FacetsGrid to App
